Third deliverable

+++ domain/model/pokemon.go
@@ -2,7 +2,7 @@ package model
 
 // Pokemon basic information
 type Pokemon struct {
-	Id   uint64 `json:"pokemon_id"`
+	Id   int    `json:"pokemon_id"`
 	Name string `json:"name"`
 	Type string `json:"type"`
 }
